requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0
